requirements.txt
langchain-gigachat  # Работа с API GigaChat для обработки запросов через LangChain
langchain-community  # Дополнительные утилиты и инструменты от сообщества LangChain
langgraph  # Построение цепочек взаимодействий между агентами и данными
langgraph-checkpoint  # Сохранение и восстановление состояния диалога
langgraph-cli  # CLI для управления агентами и цепочками через командную строку
langgraph-sdk  # Инструменты для разработки кастомных агентов и цепочек
langchain_chroma  # Интеграция LangChain с векторными базами данных Chroma
rapidfuzz  # Быстрый поиск похожих строк для обработки запросов